Add grace periods
Grace periods set an upper limit on charge capacity, while power
remains below a given threshold.

This is necessary as some inverters like the SK-SU5000E can take
in more PV power than they can output to the grid. In order to avoid
this power being lost, it should be diverted to the batteries.

To enable:
In the BatteryControl config for the affected inverters, add the
grace-capacity and grace-power-threshold properties.
In the BatteryControl time period, set grace = true if capacity
should be reserved in that period.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>